requirements.txt
Flask==2.0.2
gunicorn==20.1.0
requests==2.25.1
python-decouple==3.8
Werkzeug==2.0.1
